# internal/http/controllers/userController.go
package controllers

import (
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/spksupakorn/go-restful-authentication/internal/dto"
	"github.com/spksupakorn/go-restful-authentication/internal/http/custom"
	"github.com/spksupakorn/go-restful-authentication/internal/pkg/validator"
	"github.com/spksupakorn/go-restful-authentication/internal/usecases"

	"go.uber.org/zap"
)

// UserController handles user-related HTTP requests
type UserController struct {
	userUseCase usecases.UserUseCase
	validator   *validator.Validator
	logger      *zap.Logger
}

// NewUserController creates a new user controller
func NewUserController(
	userUseCase usecases.UserUseCase,
	validator *validator.Validator,
	logger *zap.Logger,
) *UserController {
	return &UserController{
		userUseCase: userUseCase,
		validator:   validator,
		logger:      logger,
	}
}

// Register godoc
// @Summary Register a new user
// @Description Register a new user with email and password
// @Tags auth
// @Accept json
// @Produce json
// @Param request body dto.RegisterRequest true "Registration request"
// @Success 201 {object} dto.AuthResponse
// @Failure 400 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /auth/register [post]
func (c *UserController) Register(ctx *gin.Context) {
	defer custom.PanicController(ctx)

	var req dto.RegisterRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		c.logger.Warn("Invalid request body", zap.Error(err))
		custom.PanicException(custom.NewBadRequestError("Invalid request body"))
		return
	}

	// Validate request
	if err := c.validator.Validate(&req); err != nil {
		custom.PanicException(err)
		return
	}

	// Call use case
	response, err := c.userUseCase.Register(ctx.Request.Context(), &req)
	if err != nil {
		c.logger.Error("Failed to register user", zap.Error(err))
		custom.PanicException(err)
		return
	}

	ctx.JSON(http.StatusCreated, custom.BuildResponse(custom.Success, response))
}

// Login godoc
// @Summary Login user
// @Description Authenticate user and return JWT tokens
// @Tags auth
// @Accept json
// @Produce json
// @Param request body dto.LoginRequest true "Login request"
// @Success 200 {object} dto.AuthResponse
// @Failure 400 {object} dto.ErrorResponse
// @Failure 401 {object} dto.ErrorResponse
// @Router /auth/login [post]
func (c *UserController) Login(ctx *gin.Context) {
	defer custom.PanicController(ctx)

	var req dto.LoginRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		c.logger.Warn("Invalid request body", zap.Error(err))
		custom.PanicException(custom.NewBadRequestError("Invalid request body"))
		return
	}

	// Validate request
	if err := c.validator.Validate(&req); err != nil {
		custom.PanicException(err)
		return
	}

	// Call use case
	response, err := c.userUseCase.Login(ctx.Request.Context(), &req)
	if err != nil {
		custom.PanicException(err)
		return
	}

	ctx.JSON(http.StatusOK, custom.BuildResponse(custom.Success, response))
}

// GetUserByID godoc
// @Summary Get user by ID
// @Description Get user details by ID
// @Tags users
// @Accept json
// @Produce json
// @Param id path string true "User ID"
// @Success 200 {object} dto.UserResponse
// @Failure 400 {object} dto.ErrorResponse
// @Failure 404 {object} dto.ErrorResponse
// @Security BearerAuth
// @Router /users/{id} [get]
func (c *UserController) GetUserByID(ctx *gin.Context) {
	defer custom.PanicController(ctx)

	id := ctx.Param("id")

	response, err := c.userUseCase.GetUserByID(ctx.Request.Context(), id)
	if err != nil {
		c.logger.Error("Failed to get user", zap.Error(err))
		custom.PanicException(err)
		return
	}

	ctx.JSON(http.StatusOK, custom.BuildResponse(custom.Success, response))
}

// GetAllUsers godoc
// @Summary Get all users
// @Description Get all users with pagination
// @Tags users
// @Accept json
// @Produce json
// @Param page query int false "Page number" default(1)
// @Param page_size query int false "Page size" default(10)
// @Success 200 {object} dto.UsersListResponse
// @Failure 500 {object} dto.ErrorResponse
// @Security BearerAuth
// @Router /users [get]
func (c *UserController) GetAllUsers(ctx *gin.Context) {
	defer custom.PanicController(ctx)

	page, _ := strconv.Atoi(ctx.DefaultQuery("page", "1"))
	pageSize, _ := strconv.Atoi(ctx.DefaultQuery("page_size", "10"))

	response, err := c.userUseCase.GetAllUsers(ctx.Request.Context(), page, pageSize)
	if err != nil {
		c.logger.Error("Failed to get all users", zap.Error(err))
		custom.PanicException(err)
		return
	}

	ctx.JSON(http.StatusOK, custom.BuildResponse(custom.Success, response))
}

// UpdateUser godoc
// @Summary Update user
// @Description Update user details
// @Tags users
// @Accept json
// @Produce json
// @Param id path string true "User ID"
// @Param request body dto.UpdateUserRequest true "Update request"
// @Success 200 {object} dto.UserResponse
// @Failure 400 {object} dto.ErrorResponse
// @Failure 404 {object} dto.ErrorResponse
// @Security BearerAuth
// @Router /users/{id} [put]
func (c *UserController) UpdateUser(ctx *gin.Context) {
	defer custom.PanicController(ctx)

	id := ctx.Param("id")

	var req dto.UpdateUserRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		c.logger.Warn("Invalid request body", zap.Error(err))
		custom.PanicException(custom.NewBadRequestError("Invalid request body"))
		return
	}

	// Validate request
	if err := c.validator.Validate(&req); err != nil {
		custom.PanicException(err)
		return
	}

	response, err := c.userUseCase.UpdateUser(ctx.Request.Context(), id, &req)
	if err != nil {
		c.logger.Error("Failed to update user", zap.Error(err))
		custom.PanicException(err)
		return
	}

	ctx.JSON(http.StatusOK, custom.BuildResponse(custom.Success, response))
}

// DeleteUser godoc
// @Summary Delete user
// @Description Delete user by ID
// @Tags users
// @Accept json
// @Produce json
// @Param id path string true "User ID"
// @Success 200 {object} dto.SuccessResponse
// @Failure 400 {object} dto.ErrorResponse
// @Failure 404 {object} dto.ErrorResponse
// @Security BearerAuth
// @Router /users/{id} [delete]
func (c *UserController) DeleteUser(ctx *gin.Context) {
	defer custom.PanicController(ctx)

	id := ctx.Param("id")

	err := c.userUseCase.DeleteUser(ctx.Request.Context(), id)
	if err != nil {
		c.logger.Error("Failed to delete user", zap.Error(err))
		custom.PanicException(err)
		return
	}

	ctx.JSON(http.StatusOK, custom.BuildResponse_(false, "User deleted successfully", custom.Null()))
}

// RefreshToken godoc
// @Summary Refresh access token
// @Description Generate new access token from refresh token
// @Tags auth
// @Accept json
// @Produce json
// @Param request body dto.RefreshTokenRequest true "Refresh token request"
// @Success 200 {object} map[string]string
// @Failure 400 {object} dto.ErrorResponse
// @Failure 401 {object} dto.ErrorResponse
// @Router /auth/refresh [post]
func (c *UserController) RefreshToken(ctx *gin.Context) {
	defer custom.PanicController(ctx)

	var req dto.RefreshTokenRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		c.logger.Warn("Invalid request body", zap.Error(err))
		custom.PanicException(custom.NewBadRequestError("Invalid request body"))
		return
	}

	// Validate request
	if err := c.validator.Validate(&req); err != nil {
		custom.PanicException(err)
		return
	}

	accessToken, err := c.userUseCase.RefreshToken(ctx.Request.Context(), req.RefreshToken)
	if err != nil {
		custom.PanicException(err)
		return
	}

	ctx.JSON(http.StatusOK, custom.BuildResponse_(false, "Token refreshed successfully", gin.H{
		"access_token": accessToken,
	}))
}